Fix: Cursor now works in tablet mode
While using a tablet the cursor did not change back from
the dock widget splitter cursor, after hovering over the
resize area. This bug is gone with Qt 5.12 - unfortunately
5.12 introduces a regression: mouse events are triggered
even though they should not be, while a tablet device is
active. We work around this with a custom QApplication
event handler implementation, which emits a tablet active
event to our scene view. If this gets fixed remove all of
this hacky machinery!