Feature: Add possibility to draw tool inspectors
This is implemented in a very hacky way, but it works for now and I
don't want to spend more time than necessary right now to refactor
everything to something saner.

The problem is that currently every EditorItem, which is a renderable
object in the EditorScene, owns a seperate instance of the currently
active tool, and events are only forwarded to the active item/layer.
That means that whenever the active layer changes we must remove the
current tool inspector and readd the inspector of the newly active
layers tool. This works but sounds stupid... it may be a better idea to
have a single tool active on the scene and forward the canvas of the
active layer to the tool itself, but that's probably a lot of work to
refactor - should investigate this at some point later.