Projects: Add project files instead of folders
With this change it is now possible to configure what
subfolders are contained in a data directory, for wich
layers are shown. It's also possible to configure custom
colors for each layer, as well as whether it is as mask
or a normal image layer.
This settings are saved to a segmate project file (*.spf),
which is a simple json file. Paths in this project file are
absolute and therefore not portable between systems.