Add ability to dynamically load plugins
Plugins are dynamically loaded from "~/.local/share/segmate/plugins".
A plugin is simply a python package, which exposes a class that inherits
from EditorTool in its __init__ file. Normal python import rules apply.

The plugin can expose a "__plugin_title__" variable in its __init__
file, which will be the name of the corresponding entry in the plugins
menu. If the variable does not exist, the name will simply be the name
of the package-folder.

In theory all segmate packages and modules can be imported by a plugin,
but only the "segmate.editor.editortool" module, the "segmate.editor.widgets"
package, as well as the "segmate.util" package is intended to be part
of the public API.