Don't eat 'v' key in layer widget
When pressing 'v' a set_opacity event is fired in the layer widget, even
when in a new project dialog. This is a bandaid fix: instead of
accepting the input key when the visibility toggle hotkey 'v' is pressed
we simply pass it through. Also for some reason the layer widget still
exists in the UI tree, even when the project is closed, therefore
introduce a null check.
TODO: This is pretty hacky, redo at some point.