Rework the plugin mechanism
Plugins must include a "plugin.json" file which contains the name and
entry point of an EditorTool descendent. Optionally a plugin can specify
a list of dependencies that must be installed. If a plugin is missing on
the system an installer pops up on application start, that automatically
installs all given packages in the containing virtual environment or as
"--user" if no virtual environment is present.

A minimal plugin, that depends on torch might look like this:

>> tree ~/.local/share/segmate/plugins/mytorchplugin
mytorchplugin
├── mymodule.py
└── plugin.json

>> cat plugin.json
{
    "name": "My Torch Plugin",
    "module": "mymodule",
    "class": "MyTorchTool",
    "dependencies": ["torch"]
}

>> cat mymodule.py
import torch
from segmate.editor.editortool import EditorTool

class MyTorchTool(EditorTool):

    def on_show(self):
        print("Showing MyTorchTool")