Refactor selection rendering and API
Instead of rendering a selection rectangle on top of a layer we now use
QRubberBand parented to the QGraphcisView. This ensures that the
selection rectangle is always on top and in high resolution, which looks
a lot better than the previous approach. Also refactored the API for
tools, making it more flexible and less hacky.